Skip packlist.t if List::MoreUtils is a vendor-pkg...
Many distros (eg, RedHat and Debian) provide List::MoreUtils as
a vendor-package and don't provide a packlist.  This causes the
packlist.t test to fail, interrupting many install suites.
Added logic detects such an installation (ie, an auto/List/MoreUtils
path exists under a perl @INC path, but it does not contain a
.packlist) and skips the tests.  This behaviour is analogous to
a force install, which is what is manually required currently in
this scenario.